Create a ViewInBlockExplorer component
Make a `ViewTransactionLink` component more generic/abstract in order to
use it for a type other than transaction.
